close to the end of initial version...maybe

+++ edit_data.m
@@ -5,10 +5,10 @@ function [images_train, images_test] = edit_data( images_train_orig, images_test
 
 % center images by subtracting average	
 mean = sum( images_train_orig, 2 ) / size( images_train_orig, 2 );
-images_train = images_train_orig - repmat( mean, [ 1, size( images_train_orig, 2 ) ] );
+images_train = double(images_train_orig) - repmat( mean, [ 1, size( images_train_orig, 2 ) ] );
 
 % center test data with train means
-images_test = images_test_orig - repmat( mean, [ 1, size( images_test_orig, 2 ) ] );
+images_test = double(images_test_orig) - repmat( mean, [ 1, size( images_test_orig, 2 ) ] );
